+++ benchmarks/TraceCaptureOverheadBenchmark.cs
@@ -1,11 +1,12 @@
+using System.Linq;
+
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
 
-using System.Linq;
+using Veggerby.Boards.Artifacts.Relations;
 using Veggerby.Boards.Chess;
 using Veggerby.Boards.Flows.Events;
 using Veggerby.Boards.Internal;
-using Veggerby.Boards.Artifacts.Relations;
 using Veggerby.Boards.States;
 
 namespace Veggerby.Boards.Benchmarks;